+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added Demographics Questionnaire to Onboarding and Profile [#2465](https://github.com/rokwire/illinois-app/issues/2465).
 
 ## [4.2.4] - 2022-10-17
 ### Changed